Create connected component of a graph.cpp
A graph is said to be disconnected if it is not connected, i.e. if two nodes exist in the graph such that there is no edge in between those nodes. In an undirected graph, a connected component is a set of vertices in a graph that are linked to each other by paths.

Time complexity is O( V+E ) , when implemented using the adjacency list.